+++ mip/mkext.c
@@ -298,7 +298,7 @@ main(int argc, char *argv[])
 	if (breg > mx) mx = breg;
 	if (creg > mx) mx = creg;
 	if (dreg > mx) mx = dreg;
-	if (mx > (sizeof(int)*8)-1) {
+	if (mx > (int)(sizeof(int)*8)-1) {
 		printf("too many regs in a class, use two classes instead\n");
 		printf("%d > %zu\n", mx, (sizeof(int)*8)-1);
 		rval++;
